Seperate range __repr__ for version constraint
strings.Builder is supported since v1.10,
so split files for older versions
