+++ py/range.go
@@ -6,10 +6,6 @@
 
 package py
 
-import (
-	"strings"
-)
-
 // A python Range object
 // FIXME one day support BigInts too!
 type Range struct {
@@ -110,34 +106,6 @@ func (r *Range) M__iter__() (Object, error) {
 
 func (r *Range) M__str__() (Object, error) {
 	return r.M__repr__()
-}
-
-func (r *Range) repr() (Object, error) {
-	var b strings.Builder
-	b.WriteString("range(")
-	start, err := ReprAsString(r.Start)
-	if err != nil {
-		return nil, err
-	}
-	stop, err := ReprAsString(r.Stop)
-	if err != nil {
-		return nil, err
-	}
-	b.WriteString(start)
-	b.WriteString(", ")
-	b.WriteString(stop)
-
-	if r.Step != 1 {
-		step, err := ReprAsString(r.Step)
-		if err != nil {
-			return nil, err
-		}
-		b.WriteString(", ")
-		b.WriteString(step)
-	}
-	b.WriteString(")")
-
-	return String(b.String()), nil
 }
 
 func (r *Range) M__repr__() (Object, error) {
